% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_contributions.R
\name{calc_stats_contributions_wiki}
\alias{calc_stats_contributions_wiki}
\title{Generates a summary of wiki contributions}
\usage{
calc_stats_contributions_wiki(df)
}
\arguments{
\item{df}{a dataframe with all wiki contributions generated from get_wiki_contributions()}
}
\value{
a list containing several summarised data.
}
\description{
Generates a summary of wiki contributions
}
