% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_classifications.R
\name{categorise_keys}
\alias{categorise_keys}
\title{Categorises OSM's keys}
\usage{
categorise_keys(df)
}
\arguments{
\item{df}{a dataframe with keys and values generated by \code{\link[=extract_and_combine_tags]{extract_and_combine_tags()}}}
}
\value{
a new dataframe with the same columns and two extra ones called \code{top_key} and \code{parent_key} with values defined by this function.
}
\description{
Generates a hierarchical categorisation of OSM keys, based on the type of information they convey.
For a given dataframe, this function generates two extra columns called \code{top_key} and \code{parent_key} whose values will depend on the values from the this function.
}
\examples{
df <- data.frame(
 key = c("wheelchair", "building:levels", "highway")
)
categorise_keys(df)

}
