% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create a dashboard template}
\usage{
create_dashboard(path, overwrite = FALSE)
}
\arguments{
\item{path}{a string containing the destination path where the dashboard is
going to reside.}

\item{overwrite}{boolean specifying whether to overwrite the folder if it
exists.}
}
\value{
a folder containing the template to start editing the dashboard.
}
\description{
Scaffolds the files and folders needed to produce OSM dashboards by copying
the necessary structure from inst/templates/ located in this package.
}
\examples{
# Save the current working directory
original_wd <- getwd()

# Create a temporary directory
temp_dir <- tempdir()

# Set the working directory to the temporary directory
setwd(temp_dir)

# Run the scaffold_dashboard function to create the dashboard
create_dashboard("dashboard_folder")

# Check the contents of the created dashboard folder
list.files("dashboard_folder", recursive = TRUE)

# Clean up: delete the created dashboard folder
unlink("dashboard_folder", recursive = TRUE)

# Restore the original working directory
setwd(original_wd)
}
