% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_contributions.R
\name{extract_and_combine_tags}
\alias{extract_and_combine_tags}
\title{Extract nested keys and values from changesets}
\usage{
extract_and_combine_tags(df)
}
\arguments{
\item{df}{a dataframe generated by
\code{\link[=get_contributions_changesets]{get_contributions_changesets()}}. It expects}
}
\value{
a dataframe containing all keys and values used in every changeset.
}
\description{
\code{\link[=get_contributions_changesets]{get_contributions_changesets()}} returns a dataframe with a column \code{tags}
that contains nested dataframes. This function extracts the embedded
dataframes while preserving the original changeset id.
}
