% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_contributions.R
\name{get_changesets_details}
\alias{get_changesets_details}
\title{Retrieve Changesets' details.}
\usage{
get_changesets_details(changeset_ids)
}
\arguments{
\item{changeset_ids}{a vector containing changesets ids to be queried.}
}
\value{
a dataframe describing all changes associated with the changesets.
}
\description{
Retrieves all the details from a list of changesets ids by querying
\code{\link[osmapiR:osm_download_changeset]{osmapiR::osm_download_changeset()}} for each changeset and then combining all
the results in a single dataframe.
}
