% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_contributions.R
\name{get_contributions_changesets}
\alias{get_contributions_changesets}
\title{Retrieves changesets by a set of user(s)}
\usage{
get_contributions_changesets(users, n_changesets = 100)
}
\arguments{
\item{users}{a character vector containing OSM usernames.}

\item{n_changesets}{a number specifying the maximum number of changesets to
be retrieved per user, starting from the most recent ones.}
}
\value{
a dataframe containing all changesets' details authored by the
selected users.
}
\description{
Parses a list of usernames and queries OSM api using
\code{\link[osmapiR:osm_query_changesets]{osmapiR::osm_query_changesets()}} to retrieve all the individual changesets
and then combines the results in a single dataframe.
}
