% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaries_contributions.R
\name{get_contributions_diaries}
\alias{get_contributions_diaries}
\title{Retrieve Users' diaries' titles}
\usage{
get_contributions_diaries(users)
}
\arguments{
\item{users}{a character vector containing the usernames to be queried.}
}
\value{
a dataframe containing the titles and urls of OSM diaries and
entries, per user.
}
\description{
Queries and parses the OSM diaries' RSS feeds into a dataframe.
}
