% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_contributions.R
\name{get_contributions_osm_users}
\alias{get_contributions_osm_users}
\title{Get OSM contributions by type}
\usage{
get_contributions_osm_users(users)
}
\arguments{
\item{users}{a vector containing the usernames to be queried.}
}
\value{
a dataframe containing the number of contributions by type made by
user.
}
\description{
Returns a dataframe listing the number and type of contributions
by a set of users. Data is scrapped from OSM users' profiles.
}
\examples{
get_contributions_osm_users("ccamara")
}
