% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_contributions.R
\name{get_contributions_wiki}
\alias{get_contributions_wiki}
\title{Get wiki contributions from a given set of users}
\usage{
get_contributions_wiki(users)
}
\arguments{
\item{users}{a character vector containing usernames}
}
\value{
a dataframe containing all the contributions in
wiki.openstreetmap.org made by the set of users
}
\description{
Queries OSM wiki using wikimedia API via
\code{\link[WikipediR:user_contributions]{WikipediR::user_contributions()}} and returns a dataframe with all the wiki
contributions made by every user.
}
