% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_contributions.R
\name{stats_contributions_wiki}
\alias{stats_contributions_wiki}
\title{Wiki contributions summaries}
\usage{
stats_contributions_wiki(df)
}
\arguments{
\item{df}{a dataframe with all wiki contributions generated from \code{\link[=get_contributions_wiki]{get_contributions_wiki()}}}
}
\value{
a list containing several summarised data.
}
\description{
Provides a summary of contributions
}
